schedule-agent/requirements.txt
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
anthropic>=0.34.0
flask>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pytz>=2024.1
